addming migrations

+++ CodingWiki.Models/Models/Book.cs
@@ -6,7 +6,8 @@
 
         public string Title { get; set; }
         public string ISBN { get; set; }
-        public double Price { get; set; }
+        
+        public decimal Price { get; set; }
 
     }
 }
